bootutil: Add "Swap scratch mode with status partition"
Swap scratch mode with status partition is implemented to facilitate a usage of swap mode on the devices with
a large write size of the internal and external flash memory (512B, 1024B, 4096B, 256KB, ...).
Unlike original implementation of swap scratch mode, where a status info is stored in the image trailer,
this mode introduces a special status partition in the flash memory to store a status information
about the image sectors while swapping, that guarantees a power/reset safe upgrade operation.

Signed-off-by: Oleksandr Shkurchenko <oleksandr.shkurchenko@infineon.com>